+++ docs/guides/release-notes.md
@@ -16,12 +16,16 @@ Track all ZPA Terraform provider's releases. New resources, features, and bug fi
 
 ---
 
-## 3.1.0 (December, 10 2023)
+## 3.1.0 (December, 10 2023) - Unreleased
 
 ### Notes
 
 - Release date: **(December, 10 2023)**
 - Supported Terraform version: **v1.x**
+
+### Enhacements
+
+- [PR #394](https://github.com/zscaler/terraform-provider-zpa/pull/394) - ✨ Added support for ZPA Certificate provisioning
 
 ### Fixes